Install package dependencies
Although this package was _defining_ that it needs `linter` in the
`package.json`, it was never actually calling `atom-package-deps` to
check that it was installed. Start doing this in an idle callback so it
doesn't hit the activation time of the package, and also oppotunisticly
load `atom-linter` in this callback.